requirements.txt
black
flake8
flake8-copyright
isort
matplotlib>=3.2.1
nox>=2019.11.9
numpy>=1.18.2
torch>=1.4.0
torchvision>=0.5.0
pytorch-lightning
hydra-core>=1.0.3
omegaconf>=2.0.2
pandas>=1.0.3
torchmetrics
opencv-python
boto3
onnx
onnxruntime
